==> 1_data_collection.ipynb <==
import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import warnings
warnings.filterwarnings('ignore')
import os
import shutil
import zipfile

# Create the .kaggle directory if it doesn't exist
os.makedirs(os.path.expanduser('~/.kaggle'), exist_ok=True)

# Copy the kaggle.json file to the .kaggle directory
shutil.copy('kaggle.json', os.path.expanduser('~/.kaggle/'))

# Set the permissions of the file
os.chmod(os.path.expanduser('~/.kaggle/kaggle.json'), 0o600)

!kaggle datasets download -d mfaisalqureshi/spam-email

with zipfile.ZipFile("spam-email.zip", 'r') as zip_ref:
    zip_ref.extractall("spam-email")

!kaggle datasets download -d venky73/spam-mails-dataset

with zipfile.ZipFile("spam-mails-dataset.zip", 'r') as zip_ref:
    zip_ref.extractall("spam-mails-dataset")

!kaggle datasets download -d ozlerhakan/spam-or-not-spam-dataset

with zipfile.ZipFile("spam-or-not-spam-dataset.zip", 'r') as zip_ref:
    zip_ref.extractall("spam-or-not-spam-dataset")

!kaggle datasets download -d rtatman/deceptive-opinion-spam-corpus

with zipfile.ZipFile("deceptive-opinion-spam-corpus.zip", 'r') as zip_ref:
    zip_ref.extractall("deceptive-opinion-spam-corpus")
