==> 2_merging_dataset.ipynb <==
import pandas as pd
import numpy as np

df2=pd.read_csv('spam-email\spam.csv')

df2.sample(4)

df2.rename(columns={'Category':'label','Message':'text'},inplace =True)

df2.sample(5)

df2.shape

df3=pd.read_csv('spam-mails-dataset\spam_ham_dataset.csv')

df3.sample(5)

df3.drop(columns=['label_num','Unnamed: 0'],inplace=True)

df3.sample(5)

df4=pd.read_csv('spam-or-not-spam-dataset\spam_or_not_spam.csv')

df4.sample(5)

df4['label'] = df4['label'].replace(0, 'ham')
df4['label'] = df4['label'].replace(1, 'spam')

df4=df4[['label','email']]

df4.rename(columns={'email':'text'},inplace=True)

df4.sample(5)

df5=pd.read_csv('deceptive-opinion-spam-corpus\deceptive-opinion.csv')

df5.sample(5)

df5.drop(columns=['hotel','deceptive','source'],inplace=True)
df5['polarity'] = df5['polarity'].replace('negative', 'ham')
df5['polarity'] = df5['polarity'].replace('positive', 'spam')
df5.rename(columns={'polarity':'label'},inplace=True)

df5.sample(5)

df2.shape

df3.shape

df4.shape

df5.shape

df = pd.concat([df2, df3, df4, df5], axis=0, ignore_index=True)

df.sample(5)

df.shape

df.isnull().sum()

df.dropna()

df[df.duplicated()]

df=df.drop_duplicates()

df.shape

df.to_csv('spam_notSpam.csv',index=False)
